BUG: CLI fixes etc
