+++ setup.py
@@ -42,7 +42,7 @@ setup(
         "pyyaml>=6.0",
         "click>=8.1.3",
         "metasnek>=0.0.5",
-        "snaketool-utils>=0.0.3",
+        "snaketool-utils>=0.0.4",
     ],
     entry_points={
         "console_scripts": [
